back_end/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
